Move helper method from TextureGLSingle to Texture
